+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.1.5] - 2025-09-20
+### Added
+- talent pool options
+
 ## [0.1.4] - 2025-09-18
 ### Added
 - printable report